gamelib

+++ source/SXGUIWinApi/SXGUI.cpp
@@ -28,7 +28,7 @@ See the license in LICENSE
 
 ISXGUIBaseWnd* SXGUICrBaseWnd(
 	const char* class_name, const char* caption, const char* menu,
-	WORD id, WORD x, WORD y, WORD width, WORD heigth,
+	WORD id, int x, int y, WORD width, WORD heigth,
 	HICON icon, HCURSOR cursor, HBRUSH brush,
 	DWORD exstyle, DWORD wndstyle, DWORD style,
 	HWND parent, WNDPROC handler
